+++ frontend/src/hook/usePostContextToDB.tsx
@@ -1,0 +1,9 @@
+import { useState } from "react";
+
+const usePostContextToDB = () = {
+    const [threadContents,setThreadContents] = useState("")
+
+    return threadContents
+}
+
+export default usePostContextToDB;
